refactor: Rename User to Player
