+++ spec/player_spec.rb
@@ -1,13 +1,13 @@
 require_relative 'spec_helper'
 
-RSpec.describe User do
-  subject(:user) { described_class.new }
+RSpec.describe Player do
+  subject(:player) { described_class.new }
 
   describe '#initialize' do
     it { is_expected.to be_instance_of described_class }
 
     it 'initializes with a new board' do
-      expect(user.board).to be_instance_of Board
+      expect(player.board).to be_instance_of Board
     end
   end
 end
